// src/pagesTypes/news.types.ts
export enum NewsTags {
  Latest = 'Latest',
  InFocus = 'In focus',
  MostRead = 'Most read',
  Oldest = 'Oldest',
}

export enum NewsTagsUk {
  Latest = 'Останні',
  InFocus = 'У фокусі',
  MostRead = 'Популярніші',
  Oldest = 'Старіші',
}

export enum FilterItemTopicId {
  All = 'selectTopic',
  CooperationWithBusiness = 'cooperationWithBusiness',
  EconomicRevival = 'economicRevival',
  Culture = 'culture',
  TheClubOfFriendsOfUkraine = 'theClubOfFriendsOfUkraine',
  Media = 'media',
  Medical = 'medical',
  CooperationWithOrganizationsAndFoundations = 'cooperationWithOrganizationsAndFoundations',
  InternationalAffairs = 'internationalAffairs'
}

export enum FilterItemProjectId {
  All = 'selectProject',
  TwinCitiesAndRegions = 'twinCitiesAndRegions',
  CrossMinistryRelations = 'crossMinistryRelations'
}

export enum FilterItemTagId {
  Latest = 'latest',
  InFocus = 'inFocus',
  MostRead = 'mostRead',
  Oldest = 'oldest'
}

export enum FilterItemFormatId {
  All = 'selectFormat',
  Campaign = 'campaign',
  Interview = 'interview',
  Meeting = 'meeting',
  Event = 'event',
  Report = 'report',
  Story = 'story',
  InTheMedia = 'inTheMedia',
  ProjectUpdate = 'projectUpdate',
  Speech = 'speech',
  Statement = 'statement',
}

export enum NewsTopics {
  All = '--Select topic--',
  CooperationWithBusiness = 'Cooperation with Business',
  InternationalAffairs = 'International affairs',
  EconomicRevival = 'Economic Revival',
  Culture = 'Culture',
  TheClubOfFriendsOfUkraine = 'The club of friends of Ukraine',
  Media = 'Media',
  Medical = 'Medical',
  CooperationWithOrganizationsAndFoundations = 'Cooperation with organizations and foundations',
}

export enum NewsTopicsUk {
  All = '--Обрати топік--',
  CooperationWithBusiness = 'Співпраця з бізнесом',
  EconomicRevival = 'Економічне відродження',
  Culture = 'Культура',
  TheClubOfFriendsOfUkraine = 'Клуб друзів України',
  Media = 'Медія',
  Medical = 'Медична допомога',
  CooperationWithOrganizationsAndFoundations = 'Співпраця з організаціями та фондами',
  InternationalAffairs = 'Міжнародні відносини'
}

export enum NewsProjects {
  All = '--Select project--',
  TwinCitiesAndRegions = 'Twin cities & regions',
  CrossMinistryRelations = 'Cross-ministry relations',
}

export enum NewsProjectsUk {
  All = '--Обрати проект--',
  TwinCitiesAndRegions = 'Двоміські міста та регіони',
  CrossMinistryRelations = 'Міжміністерські відносини',
}

export enum NewsFormats {
  All = '--Select format--',
  Campaign = 'Campaign',
  Interview = 'Interview',
  Meeting = 'Meeting',
  Event = 'Event',
  Report = 'Report',
  Story = 'Story',
  InTheMedia = 'In the media',
  ProjectUpdate = 'Project update',
  Speech = 'Speech',
  Statement = 'Statement'
}

export enum NewsFormatsUk {
  All = '--Обрати формат--',
  Campaign = 'Кампанія',
  Interview = 'Інтерв’ю',
  Meeting = 'Зустріч',
  Event = 'Подія',
  Report = 'Звіт',
  Story = 'Історія',
  InTheMedia = 'У ЗМІ',
  ProjectUpdate = 'Оновлення проекту',
  Speech = 'Промова',
  Statement = 'Заява'
}